+++ src/test/resources/ru/vyarus/gradle/plugin/quality/groovy/sample/GSample3.groovy
@@ -1,0 +1,10 @@
+package sample
+
+class GSample3 {
+
+    def foo(String bar) {
+        def res = "123" + bar + "123";
+        if (bar)
+            return "123"
+    }
+}