+++ CHANGELOG.md
@@ -4,9 +4,13 @@
 * Update pmd config:
     - Disable [NullAssignment](https://pmd.github.io/latest/pmd-java/rules/java/controversial.html#NullAssignment)
     check due to false positives
-* Add unified exclusions configuration: exclude extension property accepting glob values (#2)
+* Add unified exclusion patterns (glob) configuration: exclude (#2)
     - Does not affect animalsniffer because it's a different type of check
     - Findbugs did not support direct exclusions, so plugin resolves excluded classes and adds them to exclusion xml (default or user defined)
+* Add configuration to directly exclude source files from check (for exceptional cases when pattern exclusion cant help): excludeSources
+    - Does not affect animalsniffer
+    - For findbugs excluded classes will be added to exclusions xml filter
+    
 
 ### 2.1.0 (2017-01-24)
 * Update checkstyle 7.1 -> 7.4